refactor: decompose MergeEditorApp.tsx into focused modules
Extract from the 1,477-line monolith into:
- icons.tsx (125 lines): SVG icon components
- wordDiff.ts (181 lines): pure word-level diff algorithms
- mergeState.ts (112 lines): reducer, resolution helpers
- segments.tsx (523 lines): section components, code blocks, overview rail

MergeEditorApp.tsx reduced to 600 lines (App orchestrator only).
All files under the 800-line threshold.
