+++ src/webviews/react/merge-editor/mergeState.ts
@@ -1,0 +1,112 @@
+// State management and resolution helpers for the merge editor.
+// Contains the reducer, conflict resolution logic, and result builder.
+
+import type {
+    MergeEditorData,
+    MergeSegment,
+    ConflictSegment,
+    HunkResolution,
+} from "./types";
+
+export interface State {
+    data: MergeEditorData | null;
+    error: string | null;
+    resolutions: Record<number, HunkResolution>;
+}
+
+export type Action =
+    | { type: "SET_DATA"; data: MergeEditorData }
+    | { type: "SET_ERROR"; message: string }
+    | { type: "RESOLVE_HUNK"; id: number; resolution: HunkResolution };
+
+export function reducer(state: State, action: Action): State {
+    switch (action.type) {
+        case "SET_DATA":
+            return { ...state, data: action.data, error: null, resolutions: {} };
+        case "SET_ERROR":
+            return { ...state, error: action.message };
+        case "RESOLVE_HUNK":
+            return {
+                ...state,
+                resolutions: { ...state.resolutions, [action.id]: action.resolution },
+            };
+        default:
+            return state;
+    }
+}
+
+export function getResultLines(
+    segment: ConflictSegment,
+    resolution: HunkResolution | undefined,
+): string[] {
+    switch (resolution) {
+        case "ours":
+            return segment.oursLines;
+        case "theirs":
+            return segment.theirsLines;
+        case "both":
+            return [...segment.oursLines, ...segment.theirsLines];
+        case "none":
+            return [];
+        default:
+            // Non-conflicting changes auto-resolve to the changed side
+            if (segment.changeKind === "ours-only") return segment.oursLines;
+            if (segment.changeKind === "theirs-only") return segment.theirsLines;
+            return segment.baseLines;
+    }
+}
+
+export function buildResultContent(
+    data: MergeEditorData,
+    resolutions: Record<number, HunkResolution>,
+): string {
+    const { segments } = data;
+    const lines: string[] = [];
+    for (const seg of segments) {
+        if (seg.type === "common") {
+            lines.push(...seg.lines);
+        } else {
+            lines.push(...getResultLines(seg, resolutions[seg.id]));
+        }
+    }
+    const eol = data.eol ?? "\n";
+    const joined = lines.join(eol);
+    return data.hasTrailingNewline ? joined + eol : joined;
+}
+
+export function allResolved(
+    segments: MergeSegment[],
+    resolutions: Record<number, HunkResolution>,
+): boolean {
+    return segments.every(
+        (seg) =>
+            seg.type === "common" ||
+            seg.changeKind !== "conflict" ||
+            resolutions[seg.id] !== undefined,
+    );
+}
+
+export function trueConflictCount(segments: MergeSegment[]): number {
+    return segments.filter((seg) => seg.type === "conflict" && seg.changeKind === "conflict")
+        .length;
+}
+
+export function resolvedTrueConflictCount(
+    segments: MergeSegment[],
+    resolutions: Record<number, HunkResolution>,
+): number {
+    return segments.filter(
+        (seg) =>
+            seg.type === "conflict" &&
+            seg.changeKind === "conflict" &&
+            resolutions[seg.id] !== undefined,
+    ).length;
+}
+
+export function paneChangeCount(segments: MergeSegment[], side: "ours" | "theirs"): number {
+    return segments.filter((seg) => {
+        if (seg.type !== "conflict") return false;
+        if (side === "ours") return seg.changeKind !== "theirs-only";
+        return seg.changeKind !== "ours-only";
+    }).length;
+}
